docs: integrate demos into book
