--- docs/getting_started/getting_started.md
@@ -1,14 +1,0 @@
-# Getting started (Tutorial)
-
-The following pages serve as **tutorial**,
-walking you through the very basics of using `multyscale`.
-
-Along the way, they also refer to some [topic guides](../topic_guides/topic_guides),
-which have more in-depth explanation on various concepts and design decisions.
-
-First, [install `multyscale`](./installation/)
-
-<!-- ```{toctree}
-:numbered:
-
-``` -->
