--- docs/howtos/howtos.md
@@ -1,8 +1,0 @@
-# How-to... (recipes)
-
-<!-- ```{toctree}
-:titlesonly:
-    run_(F)(L)ODOG
-    explore_LODOG_parameter
-    explore_FLODOG_parameters
-``` -->
