+++ docs/quickstart/quickstart.md
@@ -1,0 +1,14 @@
+# Quick start with `multyscale`
+
+The following pages serve as a quickstart quide
+to working with the `multyscale` library.
+They are meant to introduce the core functionalities of the library,
+and assume knowledge of and familiarity with
+the concepts of multiscale spatial filtering models.
+
+The rest of this "book" is a more detailed introduction to the topic.
+
+```{toctree}
+run_(F)(L)ODOG
+construct_new
+```
