--- docs/topic_guides/topic_guides.md
@@ -1,10 +1,0 @@
-# Explanations
-
-Models of brightness perception are numerous and varied.
-One major differentiating factor, is whether a model includes
-mechanisms to explicitly resolve brightness perception phenoma;
-or whether a model attempts to more generally model early vision such,
-that brightness perception phenomena emerge but are not explicitly modeled.
-
-A (super)family of models of the latter kind that have been particularly successful,
-are multiscale spatial filtering models, examplified in the -ODOG genus of models.
